BROWSE3 REFACTOR - refactor interface files
