+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/TypeVersionPage.tsx
@@ -9,11 +9,11 @@ import {
 } from './common/SimplePageLayout';
 import {TypeVersionCategoryChip} from './common/TypeVersionCategoryChip';
 import {UnderConstruction} from './common/UnderConstruction';
-import {useWeaveflowORMContext} from './interface/wf/context';
-import {HackyTypeTree, WFTypeVersion} from './interface/wf/types';
 import {FilterableObjectVersionsTable} from './ObjectVersionsPage';
 import {FilterableOpVersionsTable} from './OpVersionsPage';
 import {FilterableTypeVersionsTable} from './TypeVersionsPage';
+import {useWeaveflowORMContext} from './wfInterface/context';
+import {HackyTypeTree, WFTypeVersion} from './wfInterface/types';
 
 export const TypeVersionPage: React.FC<{
   entity: string;
