+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/wfInterface/types.ts
@@ -1,4 +1,4 @@
-import {Type} from '../../../../../../../core';
+import {Type} from '../../../../../../core';
 
 export interface WFProject extends ProjectOwned {
   type: (name: string) => WFType;
